chore: extract cron from dashboard index (move-only)
